Fix size type

+++ src/slash.c
@@ -401,7 +401,7 @@ static void slash_complete(struct slash *slash)
 				prefix = cmd;
 				prefixlen = strlen(prefix->name);
 			} else {
-				int new_prefixlen = slash_prefix_length(prefix->name, cmd->name);
+				size_t new_prefixlen = slash_prefix_length(prefix->name, cmd->name);
 				if (new_prefixlen < prefixlen)
 					prefixlen = new_prefixlen;
 			}
